feat: rename package to @orioro/typing
This package was originally intended for only validating input/values types but usage has shown that
the other utility methods (getType, isType) as well as the capability of extending the core set of
types became should be treated as core as well. The name `typing`, rather than `validate-type`,
better expresses that new intent.

BREAKING CHANGE: rename package to @orioro/typing